+++ frontend/src/pages/Display/Bar.js
@@ -117,7 +117,7 @@ export default class Analysis extends Component {
                       {console.log('deviceDaqRealtime:', deviceDaqRealtime)}
                       {console.log('realtimeBars:', realtimeBars)}
 
-                      <BarChart width={600} height={300} data={deviceDaqRealtime}
+                      <BarChart width={1000} height={500} data={deviceDaqRealtime}
                             margin={{top: 5, right: 30, left: 20, bottom: 5}}>
                        <CartesianGrid strokeDasharray="3 3"/>
                        <XAxis dataKey="name"/>